Update SNMP-Listener-Emailer.cs
Listens for UDP messages on port 162 (this can be changed to whatever) and then emails the message using the SMTP info placed in a separate file in C:\
For offline situations, it spools all sent data to a single text file log called the offlien buffer and then sends that as soon as it is back online. 
Simply target the other Open IT Tools at the IP of this at the same udp port and bam. instant RMM for free without any tagware or crapware or anything out of your control.